pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hatch-demo"
dynamic = ["version"]
description = 'Hatch demo for Python Rennes Meetup 04-2024'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Frédéric Collonval", email = "fcollonval@webscit.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "connexion~=3.0",
  "hatchling>=0.21.0",
  "pyyaml~=6.0",
  "uvicorn"
]

[project.urls]
Documentation = "https://github.com/unknown/hatch-demo#readme"
Issues = "https://github.com/unknown/hatch-demo/issues"
Source = "https://github.com/unknown/hatch-demo"

[project.entry-points.hatch]
openapi = "hatch_demo.hooks"

[tool.hatch.version]
source = "openapi"
path = "src/hatch_demo/spec/openapi.yaml"

[tool.hatch.build.hooks.version]
path = "src/hatch_demo/_version.py"
template = """
# This file is auto-generated by Hatchling. As such, do not:
#   - modify
#   - track in version control e.g. be sure to add to .gitignore
__version__ = '{version}'"""

[tool.hatch.envs.default]
dependencies = [
  "httpie",
  "coverage[toml]>=6.5",
  "pytest",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/hatch_demo tests}"

[tool.coverage.run]
source_pkgs = ["hatch_demo", "tests"]
branch = true
parallel = true
omit = [
  "src/hatch_demo/__about__.py",
]

[tool.coverage.paths]
hatch_demo = ["src/hatch_demo", "*/hatch-demo/src/hatch_demo"]
tests = ["tests", "*/hatch-demo/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
